fix: resolve all clippy warnings across workspace
- Add #[allow(clippy::unwrap_used, clippy::expect_used)] to test modules
- Remove needless struct update, redundant casts, identity ops
- Fix manual_strip, redundant_closure, let_and_return, module_inception
- Add clippy::disallowed_types allow for test files using serde_json::Value

+++ crates/common/src/pricing_config.rs
@@ -720,12 +720,13 @@ pub enum ValidationError {
 }
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used, clippy::expect_used)]
 mod tests {
     use super::*;
 
     /// Helper function to convert dollars to nanodollars as i64
     fn to_nano(price: f64) -> i64 {
-        dollars_to_nano(price) as i64
+        dollars_to_nano(price)
     }
 
     #[test]
